requirements.txt
Pillow==10.1.0
cairosvg==2.7.0
svgwrite==1.4.3
reportlab==4.0.7
numpy==1.24.3
pytest==7.4.3
pytest-cov==4.1.0
pytest-xdist==3.5.0
black==23.11.0
pylint==3.0.2
mypy==1.7.0
bandit==1.7.5
